requirements.txt
Flask
Flask_Bootstrap
newsapi-python
FinNews
PyYAML
pandas
datetime
statistics
dateutils
regex
lxml
py3socket
Py3Collections
gunicorn